feat(scripts): Rewrite seed script to use ClickHouse architecture
Complete rewrite of score analytics seeding script to match Langfuse's
dual-database architecture:

**Key Changes:**
- Use ClickHouse for traces, observations, and scores (not Postgres)
- Import factory functions: createTrace, createObservation, createTraceScore
- Use batch insertion: createTracesCh, createObservationsCh, createScoresCh
- Update timestamps to milliseconds (Date.now()) instead of Date objects
- Batch insertions (500 records) for better performance

**Architecture:**
- Traces/observations/scores live in ClickHouse, not Postgres
- Prisma LegacyPrisma* models are deprecated
- Factory functions provide sensible defaults and type safety
- Batch insertion functions handle ClickHouse-specific formatting

**Status:**
Script logic is correct but currently blocked by Node v24 + AWS SDK
@smithy/core dependency issue affecting entire seed infrastructure.
Documented in README.
